requirements.txt
# Core ML Dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# NLP & Text Processing
nltk==3.8.1

# UI Framework
streamlit==1.28.1

# Visualization
plotly==5.17.0
matplotlib==3.8.0

# Generative AI
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0